requirements.txt
# Core
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Machine Learning
xgboost>=2.0.0
scikit-learn>=1.4.0
shap>=0.45.0

# Visualization
plotly>=5.18.0

# Web Interface
gradio>=5.0.0

# Utilities
joblib>=1.3.0